Move summary sections back to /components, now inside SubmissionSummarySection
