+++ services/app-web/src/components/SubmissionSummarySection/ContractDetailsSummarySection/index.ts
