+++ services/app-web/src/components/SubmissionSummarySection/DocumentsSummarySection/DocumentsSummarySection.module.scss
@@ -1,6 +1,6 @@
-@import '../SubmissionSummary.module.scss';
+@import '../SubmissionSummarySection.module.scss';
 
-.reviewSection {
+.summarySection {
     ul {
         list-style-type: none;
         margin: 0;
