+++ services/app-web/src/components/SubmissionSummarySection/RateDetailsSummarySection/index.ts
