+++ services/app-web/src/components/SubmissionSummarySection/SubmissionSummarySection.module.scss
@@ -1,6 +1,5 @@
 @import '../../styles/uswdsImports.scss';
 @import '../../styles/custom.scss';
-
 .summarySectionWrapper {
     max-width: 50rem;
 }
@@ -25,13 +24,8 @@
     }
 
     ul {
-        list-style-type: none;
-        margin: 0;
+        margin: 0 0 0 units(3);
         padding: 0;
-
-        li {
-            padding: units(1) 0;
-        }
     }
 
     // align state contacts
