+++ services/app-web/src/components/SubmissionSummarySection/SubmissionTypeSummarySection/index.ts
