+++ services/app-web/src/components/SubmissionSummarySection/index.ts
@@ -1,0 +1,5 @@
+export { SubmissionTypeSummarySection } from './SubmissionTypeSummarySection'
+export { ContractDetailsSummarySection } from './ContractDetailsSummarySection'
+export { RateDetailsSummarySection } from './RateDetailsSummarySection'
+export { ContactsSummarySection } from '../../components/SubmissionSummarySection/ContactsSummarySection'
+export { DocumentsSummarySection } from './DocumentsSummarySection'
