+++ services/app-web/src/pages/SubmissionSummary/SubmissionSummary.module.scss
@@ -10,92 +10,3 @@
     max-width: 50rem; // should match ReviewSubmit max-width
     padding: 2rem 0;
 }
-
-.reviewSectionWrapper {
-    max-width: 50rem;
-}
-
-.reviewSection {
-    margin: units(2) auto;
-    background: $cms-color-white;
-    padding: units(2) units(4);
-    border: 1px solid $theme-color-base-lighter;
-    line-height: units(3);
-    @include u-radius('md');
-
-    h2 {
-        margin: 0;
-        @include u-text('normal');
-    }
-    &:first-of-type {
-        h2 {
-            @include u-text('bold');
-            font-size: size('body', 'lg');
-        }
-    }
-
-    ul {
-        margin: 0 0 0 units(3);
-        padding: 0;
-    }
-
-    // align state contacts
-    dl div[class^='grid-container'] {
-        padding: 0;
-
-        div {
-            padding-bottom: units(2);
-        }
-
-        // dont bottom pad last two grid items
-        div:nth-last-child(-n + 2) {
-            padding-bottom: 0;
-        }
-    }
-}
-
-.reviewSectionHeader {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: units(4);
-    padding-bottom: units(2);
-    border-bottom: 1px solid $theme-color-base-lighter;
-}
-
-.reviewSectionSubHeader {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: units(4) 0;
-    padding-top: units(4);
-    border-top: 1px solid $theme-color-base-lighter;
-}
-
-.contactInfo p {
-    margin: units(1) 0;
-}
-
-.documentDesc {
-    margin: 0;
-}
-
-.submitButton {
-    background: $theme-color-success;
-    &:hover {
-        background-color: #2a7a3b !important;
-    }
-}
-
-@media print {
-    header,
-    footer,
-    .pageActions,
-    .reviewSectionHeader a {
-        display: none;
-    }
-    .reviewSection {
-        border: none;
-        padding: 0;
-    }
-}
