+++ services/app-web/src/pages/SubmissionSummary/index.ts
@@ -1,7 +1,1 @@
 export { SubmissionSummary } from './SubmissionSummary'
-
-export { SubmissionTypeSummarySection } from './SubmissionTypeSummarySection'
-export { ContractDetailsSummarySection } from './ContractDetailsSummarySection'
-export { RateDetailsSummarySection } from './RateDetailsSummarySection'
-export { ContactsSummarySection } from './ContactsSummarySection'
-export { DocumentsSummarySection } from './DocumentsSummarySection'
